requirements.txt
fastapi
uvicorn
pydantic
pandas
ultralytics
opencv-python-headless
pillow
numpy
libgl1
libglib2.0-0
